upload objects classes moved from main project to upload manager library

+++ app/MediaUploader/Sitecore.Mobile.MediaUploader/sc_QuickImageViewController.h
@@ -14,7 +14,7 @@
 
 @property (nonatomic) IBOutlet SCImageView *imageView;
 @property (nonatomic, retain) NSMutableArray*  items;
-@property (nonatomic) int selectedImage;
+@property (nonatomic) NSUInteger selectedImage;
 @property (nonatomic) IBOutlet UICollectionView *collectionView;
 @property (nonatomic, strong) SCApiSession *session;
